:construction_worker: adding bell endpoints, decomposing listentry, addind drawer button
