+++ assets/translations/de.json
@@ -14,6 +14,7 @@
   },
 
   "drawer": {
+    "notifications": "Benachrichtigungen",
     "profile": "Profil",
     "logout": "Abmelden",
     "feedback": "Feedback"
